admin screen unit

+++ src/server/api/services/unit.service.ts
@@ -1,21 +1,33 @@
 import { CategorySchemaType, CategorySchemaUpdateType } from '@/libs/schema/category.schema'
 import { prisma } from '@/server/db'
+import { TRPCError } from '@trpc/server'
 import { User } from 'next-auth'
 import { UtilsService } from './utils.service'
 
 class UnitService extends UtilsService {
   async getAll() {
-    const unit = await prisma.unit.findMany()
+    const unit = await prisma.unit.findMany({
+      include: {
+        Product: true,
+      },
+    })
 
     return unit
   }
 
   async get(id: string) {
-    const unit = await prisma.category.findUnique({
+    const unit = await prisma.unit.findUnique({
       where: {
         id,
       },
     })
+
+    if (!unit) {
+      throw new TRPCError({
+        code: 'NOT_FOUND',
+        message: 'error.unit-not-found',
+      })
+    }
 
     return unit
   }
@@ -27,12 +39,18 @@ class UnitService extends UtilsService {
       data,
     })
 
+    if (!unit) {
+      throw new TRPCError({
+        code: 'TIMEOUT',
+        message: 'error.create-failed',
+      })
+    }
+
     return unit
   }
 
   async update(data: CategorySchemaUpdateType, user: User) {
     await this.CheckAdmin(user.id)
-
     const { id, name } = data
 
     const unit = await prisma.unit.update({
@@ -44,12 +62,31 @@ class UnitService extends UtilsService {
       },
     })
 
+    if (!unit) {
+      throw new TRPCError({
+        code: 'NOT_FOUND',
+        message: 'error.unit-not-found',
+      })
+    }
+
     return unit
   }
 
   async delete(id: string, user: User) {
     await this.CheckAdmin(user.id)
 
+    const unit = await prisma.unit.findUnique({
+      where: {
+        id,
+      },
+    })
+
+    if (!unit) {
+      throw new TRPCError({
+        code: 'NOT_FOUND',
+        message: 'error.unit-not-found',
+      })
+    }
     await prisma.unit.delete({
       where: {
         id,
@@ -57,6 +94,23 @@ class UnitService extends UtilsService {
     })
 
     return 'Delete success'
+  }
+
+  async byId(id: string) {
+    const unit = await prisma.unit.findUnique({
+      where: {
+        id,
+      },
+    })
+
+    if (!unit) {
+      throw new TRPCError({
+        code: 'NOT_FOUND',
+        message: 'error.unit-not-found',
+      })
+    }
+
+    return unit
   }
 }
